chore: organize folders
Move form components into their own folder so we can isolate them.
