+++ apps/studio/components/interfaces/Account/AccessTokens/index.ts
@@ -2,6 +2,6 @@ export { default as AccessTokenList } from './AccessTokenList'
 export { default as NewAccessTokenButton } from './NewAccessTokenButton'
 export { default as NewTokenBanner } from './NewTokenBanner'
 export { ViewTokenPermissionsPanel } from './ViewTokenPermissionsPanel'
-export { TokenBasicInfoForm } from './TokenBasicInfoForm'
-export { TokenResourceAccessForm } from './TokenResourceAccessForm'
-export { TokenPermissionsForm } from './TokenPermissionsForm'
+export { TokenBasicInfoForm } from './Form/TokenBasicInfoForm'
+export { TokenResourceAccessForm } from './Form/TokenResourceAccessForm'
+export { TokenPermissionsForm } from './Form/TokenPermissionsForm'
